requirements.txt
numpy~=1.22.4
scikit-learn~=1.1.1
python-dateutil~=2.8.2
Flask~=2.1.2
pandas~=1.4.2
keras~=2.9.0
gunicorn